Fix a couple of crash bugs in the encoder
Tested only on an amd64 GNU/Linux system; may break builds on other
platforms.

+++ gameplay-encoder/src/Thread.h
@@ -64,7 +64,7 @@ namespace gameplay
         void* arg = data->arg;
         delete data;
         data = NULL;
-        int retVal = threadFunction(arg);
+        intptr_t retVal = threadFunction(arg);
         pthread_exit((void*)retVal);
     }
